requirements.txt
streamlit 
python-dotenv 
google-generativeai 
pdf2image
pymupdf